Placing stuff in backend_bases.py and renaming rc param to single_window
