+++ matplotlibrc.template
@@ -39,7 +39,7 @@ backend      : %(backend)s
 # If you are using one of the GTK3 backends (GTK3Agg or GTK3Cairo)
 # you can set to use only one window with tabbed figures instead of 
 # multiple windows one for each figure
-#backend.gtk3.tabbed : True
+#backend.single_window : True
 
 # Note that this can be overridden by the environment variable
 # QT_API used by Enthought Tool Suite (ETS); valid values are
